moved decrypt shards to different file, added more updates according to comments
